Rework Registration; Rename Namespaces; Add DelayedActions;\n Replace Assembly with NuGet Packege
Registration is done automatically for all items that have the proper attribute, saving space and preventing accidental double-registration.

A spelling mistake in "Ingredients" was corrected, And namespaces now match file structure for better organization.

new work-in-progress Delayed Actions, that act after a certain span of in-game time. yet untested but in theory functional.

replaces the game's assembly with the NuGet package for the GameLibs, and removed Krafs.Publicizer for easy of use